+++ pipelines/treatment__subsidio_sppo_apuracao/constants.py
@@ -25,7 +25,8 @@ WEBHOOK_KEY = "subsidio_data_check"
 ADDITIONAL_VARS = {"tipo_teste": "subsidio"}
 
 PRE_TEST_SELECT = (
-    "veiculo_dia tecnologia_servico viagem_planejada transacao transacao_riocard"
+    "sppo_registros sppo_realocacao check_gps_treatment__gps_sppo sppo_veiculo_dia"
+    " veiculo_dia tecnologia_servico viagem_planejada transacao transacao_riocard"
     " gps_validador test_completude__temperatura test_jae_captura_subsidio viagem_completa"
 )
 PRE_TEST_V25_SELECT = (
@@ -45,6 +46,33 @@ PRE_CHECKS_LIST = {
             "Captura da Jaé sem timestamps ausentes em `transacao`, `transacao_riocard`"
             " e `gps_validador`"
         )
+    },
+    "sppo_realocacao": {
+        "check_gps_capture__sppo_realocacao": {
+            "description": "Todos os dados de realocação foram capturados"
+        }
+    },
+    "sppo_registros": {
+        "check_gps_capture__sppo_registros": {
+            "description": "Todos os dados de GPS foram capturados"
+        }
+    },
+    "gps_sppo": {
+        "check_gps_treatment__gps_sppo": {
+            "description": "Todos os dados de GPS foram devidamente tratados"
+        },
+        "dbt_utils__unique_combination_of_columns__gps_sppo": {
+            "description": "Todos os registros são únicos"
+        },
+    },
+    "sppo_veiculo_dia": {
+        "not_null": {"description": "Todos os valores da coluna `{column_name}` não nulos"},
+        "dbt_utils__unique_combination_of_columns__data_id_veiculo__sppo_veiculo_dia": {
+            "description": "Todos os registros são únicos"
+        },
+        "dbt_expectations__expect_row_values_to_have_data_for_every_n_datepart__sppo_veiculo_dia": {
+            "description": "Todas as datas possuem dados"
+        },
     },
     "veiculo_dia": {
         "not_null": {"description": "Todos os valores da coluna `{column_name}` não nulos"},